add SearchGroupOrFriendActivity.java CreateGroupActivity.java and related resources

+++ src/com/trible/scontact/controller/IFriendsControl.java
@@ -2,12 +2,11 @@ package com.trible.scontact.controller;
 
 import java.util.List;
 
-import com.trible.scontact.models.Groupsinfo;
 import com.trible.scontact.models.Friendinfo;
 
 public interface IFriendsControl {
 
-	public List<Friendinfo> getFriendsListByGroupId(int groupid);
+	public List<Friendinfo> getFriendsListByGroupId(long groupid);
 	public boolean deleteFriend(int friendId);
 	public boolean createFriend(Friendinfo info);
 	public boolean updateFriend(Friendinfo uInfo);
